+++ app/src/main/java/com/robosoft/archana/recyclerviewassignment/Modal/EditFragmentCommunicator.java
@@ -6,5 +6,5 @@ import java.util.ArrayList;
  * Created by archana on 12/2/16.
  */
 public interface EditFragmentCommunicator {
-    public void toSendEdittedList(ArrayList<ProductList> arrayList,int position);
+    public void toSendEdittedList(int position);
 }
