+++ app/src/main/java/com/robosoft/archana/recyclerviewassignment/Network/ParserinInBackground.java
@@ -47,7 +47,6 @@ public class ParserinInBackground extends AsyncTask<Void, Void, ArrayList<Produc
             return null;
         }
 
-
         return produCArrayList;
     }
