+++ app/src/main/java/com/robosoft/archana/recyclerviewassignment/adapter/ProductListAdapter.java
@@ -5,25 +5,21 @@ import android.content.DialogInterface;
 import android.graphics.Bitmap;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-import android.os.Bundle;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.util.LruCache;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.animation.AnimationUtils;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.robosoft.archana.recyclerviewassignment.Modal.FragmentCommunicator;
 import com.robosoft.archana.recyclerviewassignment.Modal.ProductList;
 import com.robosoft.archana.recyclerviewassignment.Network.ImageDownloader;
 import com.robosoft.archana.recyclerviewassignment.R;
-import com.robosoft.archana.recyclerviewassignment.fragment.EditFragment;
+import com.robosoft.archana.recyclerviewassignment.fragment.ProductFragment;
 
 import java.util.ArrayList;
 
@@ -59,21 +55,20 @@ public class ProductListAdapter extends RecyclerView.Adapter<ProductListAdapter.
     public void onBindViewHolder(ViewHolder holder, final int position) {
         final ProductList productListClass = mProductList.get(position);
         holder.mTextName.setText(productListClass.getmName());
-        holder.mTextCost.setText("" + productListClass.getmCost());
+        holder.mTextCost.setText(String.valueOf(productListClass.getmCost()));
         holder.mTextDesription.setText(productListClass.getmDesription());
         if(isNetworkAvailable()){
-            ImageDownloader imageDownloader = new ImageDownloader(mLrucache, productListClass.getmImage(), holder.mImage);
-            imageDownloader.execute();
+            new ImageDownloader(mLrucache, productListClass.getmImage(), holder.mImage).execute();
         }
-        else
+       else
         {
-         holder.mImage.setImageResource(R.drawable.download);
+            holder.mImage.setImageResource(R.drawable.download);
         }
 
         holder.mButtonUpdate.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                EditFragment editFragment = new EditFragment();
+                ProductFragment editFragment = new ProductFragment();
                 mEditableList.add(productListClass);
                 fragmentCommunicator.toGoToFragment(editFragment, position, mEditableList);
             }
@@ -82,8 +77,8 @@ public class ProductListAdapter extends RecyclerView.Adapter<ProductListAdapter.
             @Override
             public void onClick(View v) {
                 AlertDialog.Builder builder = new AlertDialog.Builder(mContext);
-                builder.setTitle("Delete Alert!");
-                builder.setMessage("Are you sure want to delete?");
+                builder.setTitle(R.string.titleOfAlertDialog);
+                builder.setMessage(R.string.messageOfAlertDialog);
                 builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
@@ -101,7 +96,7 @@ public class ProductListAdapter extends RecyclerView.Adapter<ProductListAdapter.
             }
         });
 
-        //Custome animation
+    //Custome animation
         if (position > mPreviousPositon) {
             com.robosoft.archana.recyclerviewassignment.Modal.AnimationUtils.animatie(holder, true);
         } else {
@@ -110,7 +105,7 @@ public class ProductListAdapter extends RecyclerView.Adapter<ProductListAdapter.
         mPreviousPositon = position;
 
     }
-    private boolean isNetworkAvailable() {
+  private boolean isNetworkAvailable() {
         ConnectivityManager connectivityManager
                 = (ConnectivityManager)mContext.getSystemService(Context.CONNECTIVITY_SERVICE);
         NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
