+++ app/src/main/java/com/robosoft/archana/recyclerviewassignment/fragment/ProductFragment.java
@@ -4,8 +4,6 @@ package com.robosoft.archana.recyclerviewassignment.fragment;
 import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,6 +11,7 @@ import android.widget.Button;
 import android.widget.EditText;
 
 import com.robosoft.archana.recyclerviewassignment.Modal.Communicator;
+import com.robosoft.archana.recyclerviewassignment.Modal.EditFragmentCommunicator;
 import com.robosoft.archana.recyclerviewassignment.Modal.ProductList;
 import com.robosoft.archana.recyclerviewassignment.R;
 
@@ -21,16 +20,17 @@ import java.util.ArrayList;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class AddProductFragment extends Fragment {
+public class ProductFragment extends Fragment {
 
     private ArrayList<ProductList> addedProductList = new ArrayList<>();
     private Button mAddButton, mCancelButton;
     private EditText mEditName, mEditCost, mEditImage, mEditDescription;
     private View mOneRow;
     Communicator communicator;
+    EditFragmentCommunicator editFragmentCommunicator;
     private Context mContext;
 
-    public AddProductFragment() {
+    public ProductFragment() {
         // Required empty public constructor
     }
 
@@ -41,6 +41,7 @@ public class AddProductFragment extends Fragment {
         // Inflate the layout for this fragment
         mContext = container.getContext();
         communicator = (Communicator) mContext;
+        editFragmentCommunicator = (EditFragmentCommunicator) mContext;
         mOneRow = inflater.inflate(R.layout.fragment_add_product, container, false);
 
         mEditName = (EditText) mOneRow.findViewById(R.id.feditname);
@@ -49,6 +50,7 @@ public class AddProductFragment extends Fragment {
         mEditDescription = (EditText) mOneRow.findViewById(R.id.feditdesription);
 
         mAddButton = (Button) mOneRow.findViewById(R.id.fadd);
+
         mCancelButton = (Button) mOneRow.findViewById(R.id.fcancel);
 
 
@@ -67,7 +69,7 @@ public class AddProductFragment extends Fragment {
                 productList.setmDesription(description);
                 addedProductList.add(productList);
                 communicator.toCommunicate(addedProductList);
-                getActivity().getSupportFragmentManager().beginTransaction().remove(AddProductFragment.this).commit();
+                getActivity().getSupportFragmentManager().beginTransaction().remove(ProductFragment.this).commit();
             }
         });
         mCancelButton.setOnClickListener(new View.OnClickListener() {
@@ -87,8 +89,7 @@ public class AddProductFragment extends Fragment {
                     mEditDescription.setText(" ");
                 }
 
-                // communicator.toCommunicate(addedProductList);
-                //getActivity().getSupportFragmentManager().beginTransaction().remove(AddProductFragment.this).commit();
+
             }
         });
         return mOneRow;
@@ -99,7 +100,45 @@ public class AddProductFragment extends Fragment {
     @Override
     public void onActivityCreated(Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
+        mCancelButton = (Button) mOneRow.findViewById(R.id.editfragcancel);
+        Bundle bundle = getArguments();
+        if(bundle==null){
+            mAddButton.setText("Add");
+        }
+        else {
+            mAddButton.setText("Update");
+            final int position = bundle.getInt("Position");
+            final ArrayList<ProductList> editablelist = (ArrayList<ProductList>) bundle.getSerializable("EditableList");
+            for(int i = 0;i < editablelist.size();i++){
+                ProductList productList1 = editablelist.get(i);
+                String name = productList1.getmName();
+                mEditName.setText(name);
+                mEditCost.setText(String.valueOf( productList1.getmCost()));
+                String image = productList1.getmImage();
+                mEditImage.setText(image);
+                String description = productList1.getmDesription();
+                mEditDescription.setText(description);
+            }
+            final ArrayList<ProductList> arrayList = (ArrayList<ProductList>) bundle.getSerializable("List");
+            mAddButton.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    ProductList productList = arrayList.get(position);
+                    String editedname = mEditName.getText().toString();
+                    int editedcost = Integer.parseInt(mEditCost.getText().toString());
+                    String editedimage = mEditImage.getText().toString();
+                    String editeddescription = mEditDescription.getText().toString();
+                    productList.setmName(editedname);
+                    productList.setmCost(editedcost);
+                    productList.setmImage(editedimage);
+                    productList.setmDesription(editeddescription);
+                    arrayList.set(position, productList);
+                    editFragmentCommunicator.toSendEdittedList(position);
+                    getActivity().getSupportFragmentManager().beginTransaction().remove(ProductFragment.this).commit();
 
+                }
+            });
+        }
 
     }
