+++ app/src/main/res/layout/fragment_add_product.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
 
-    tools:context="com.robosoft.archana.recyclerviewassignment.fragment.AddProductFragment">
+    tools:context=".fragment.ProductFragment">
     <!-- <TextView
          android:layout_width="wrap_content"
          android:layout_centerHorizontal="true"
