+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,6 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="titleOfAlertDialog">Delete Alert!</string>
+    <string name="messageOfAlertDialog">Are you sure want to delete?</string>
 </resources>
